Start drafting how up-to-date checks will be.
This is probably where we'll differ from existing implementations. I've
no wish for dealing with a SQL database and will probably choose a NoSQL
key-value store for this. I haven't looked on how the C++ implementation
manages the hashes (or if it uses timestamps after all), but the
file-oriented solution @jekor proposed probably won't be what I use.