Add d-leveldb (a leveldb wrapper library).
I'm not sure about using leveldb as redo's dependency database. It has a
major drawback which is I can't run multiple clients on it, unless I
write a writer daemon and set-up messaging myself. This breaks the
possibility of running several redo and redo-ifchange instances
concurrently right off-the-bat.

Still, I may choose to use it and work around the issues. I'd like to
discuss this at #2.